Update random-song-guessing.html
added the point system. this point system is special :)

when the user gets a song name correctly, the user is rewarded 30 points. when the user gets the song name wrong, the user minus the amount of points equivalent to the amount of time the user took to choose an option. 

for example: if the user took 6 seconds to choose the wrong option, the user will minus 6 points from their current point. 
Also, each time the user is spent listening to the song, the points minus by 1 for each second as well. 

this point system is to pressure the user to choose an option fast.