Update random-song-guessing.html
changed the game system:
now players will start with 30 pts.

when the points reach to zero, but the user is still guessing the song, user will still be able to input the options. when the user input the wrong options, game will be over and the buttons will all be locked. user has the option to retry.